// src/app/models/result.model.ts
export interface Result {
  name: string
  height: string
  mass: string
  hair_color: string
  skin_color: string
  eye_color: string
  birth_year: string
  gender: string
  homeworld: string
  films: string[]
  species: any[]
  vehicles: string[]
  starships: any[]
  created: string
  edited: string
  url: string
}